requirements.txt
aiohttp==3.9.3
pycurl
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
pycryptodome==3.10.1
Deprecated==1.2.14
fastapi==0.109.2
frozenlist==1.4.1
idna==3.6
jwcrypto==1.5.1
lxml==5.1.0
multidict==6.0.5
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
PyPDF2==3.0.1
python-docx==1.1.0
python-jwt==4.1.0
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.0
wrapt==1.16.0
yarl==1.9.4
uvicorn==0.27.1
websockets==12.0
google-generativeai==0.3.2
python-dotenv==1.0.1
langchain-community
langchain-core
langsmith
langchain==0.2.10
langchain-text-splitters
openai
langchain-openai==0.1.17
tiktoken==0.7.0
tenacity==8.3.0
supabase==2.3.0
fastapi-limiter==0.1.6
redis==5.1.0b7